extract launch.ts
